refactor: move deps to correct place
